Seprated the main()
Some minor updates..
