+++ dice_roll.java
@@ -1,0 +1,24 @@
+//PLEASE EXPLAIN WHAT NEEDS TO BE DONE HERE!!!!!!!!!!!!!!!
+import java.util.*;
+public class dice_roll
+{
+ public static void main(String[] args){
+     Scanner input = new Scanner(System.in);
+     Luckyseven dice = new Luckyseven();
+     dice.roll();
+     dice.getDollars();
+    
+     int d =input.nextInt();
+     dice.setDollars(d);
+ 
+ 
+
+ 
+
+
+
+
+
+
+}
+}
